Added /dist to .gitignore. Updates to index.ts
I will temporarily remove /dist to make commit history readable and will bring it back later for ease of use for the project.

fetchAssets.ts removes anything not in defined extensions (.jpg...) in all images. Minor updates to index.ts to not rely on a click function, instead a general loadAssets() for any use case.